src/requirements.txt
awslambdaric==1.2.2
boto3==1.18.41
pytest==6.2.4
Pillow==8.3.1
